Reorganize components in `ui2`
