+++ crates/ui2/src/components.rs
@@ -1,71 +1,45 @@
-mod assistant_panel;
-mod breadcrumb;
-mod buffer;
-mod buffer_search;
-mod chat_panel;
-mod collab_panel;
-mod command_palette;
+mod avatar;
+mod button;
 mod context_menu;
-mod copilot;
-mod editor_pane;
+mod details;
 mod facepile;
+mod icon;
 mod icon_button;
+mod indicator;
+mod input;
 mod keybinding;
-mod language_selector;
+mod label;
 mod list;
 mod modal;
-mod multi_buffer;
 mod notification_toast;
-mod notifications_panel;
 mod palette;
 mod panel;
-mod panes;
+mod player;
 mod player_stack;
-mod project_panel;
-mod recent_projects;
-mod status_bar;
+mod stack;
 mod tab;
-mod tab_bar;
-mod terminal;
-mod theme_selector;
-mod title_bar;
 mod toast;
-mod toolbar;
-mod traffic_lights;
-mod workspace;
+mod tool_divider;
 
-pub use assistant_panel::*;
-pub use breadcrumb::*;
-pub use buffer::*;
-pub use buffer_search::*;
-pub use chat_panel::*;
-pub use collab_panel::*;
-pub use command_palette::*;
+pub use avatar::*;
+pub use button::*;
 pub use context_menu::*;
-pub use copilot::*;
-pub use editor_pane::*;
+pub use details::*;
 pub use facepile::*;
+pub use icon::*;
 pub use icon_button::*;
+pub use indicator::*;
+pub use input::*;
 pub use keybinding::*;
-pub use language_selector::*;
+pub use label::*;
 pub use list::*;
 pub use modal::*;
-pub use multi_buffer::*;
 pub use notification_toast::*;
-pub use notifications_panel::*;
 pub use palette::*;
 pub use panel::*;
-pub use panes::*;
+pub use player::*;
 pub use player_stack::*;
-pub use project_panel::*;
-pub use recent_projects::*;
-pub use status_bar::*;
+pub use stack::*;
 pub use tab::*;
-pub use tab_bar::*;
-pub use terminal::*;
-pub use theme_selector::*;
-pub use title_bar::*;
 pub use toast::*;
-pub use toolbar::*;
-pub use traffic_lights::*;
-pub use workspace::*;
+pub use tool_divider::*;
