+++ crates/ui2/src/components/avatar.rs
