+++ crates/ui2/src/components/details.rs
