+++ crates/ui2/src/components/icon.rs
