+++ crates/ui2/src/components/indicator.rs
