+++ crates/ui2/src/components/player.rs
