+++ crates/ui2/src/components/stack.rs
