--- crates/ui2/src/elements.rs
@@ -1,21 +1,0 @@
-mod avatar;
-mod button;
-mod details;
-mod icon;
-mod indicator;
-mod input;
-mod label;
-mod player;
-mod stack;
-mod tool_divider;
-
-pub use avatar::*;
-pub use button::*;
-pub use details::*;
-pub use icon::*;
-pub use indicator::*;
-pub use input::*;
-pub use label::*;
-pub use player::*;
-pub use stack::*;
-pub use tool_divider::*;
