+++ crates/ui2/src/to_extract.rs
@@ -1,0 +1,47 @@
+mod assistant_panel;
+mod breadcrumb;
+mod buffer;
+mod buffer_search;
+mod chat_panel;
+mod collab_panel;
+mod command_palette;
+mod copilot;
+mod editor_pane;
+mod language_selector;
+mod multi_buffer;
+mod notifications_panel;
+mod panes;
+mod project_panel;
+mod recent_projects;
+mod status_bar;
+mod tab_bar;
+mod terminal;
+mod theme_selector;
+mod title_bar;
+mod toolbar;
+mod traffic_lights;
+mod workspace;
+
+pub use assistant_panel::*;
+pub use breadcrumb::*;
+pub use buffer::*;
+pub use buffer_search::*;
+pub use chat_panel::*;
+pub use collab_panel::*;
+pub use command_palette::*;
+pub use copilot::*;
+pub use editor_pane::*;
+pub use language_selector::*;
+pub use multi_buffer::*;
+pub use notifications_panel::*;
+pub use panes::*;
+pub use project_panel::*;
+pub use recent_projects::*;
+pub use status_bar::*;
+pub use tab_bar::*;
+pub use terminal::*;
+pub use theme_selector::*;
+pub use title_bar::*;
+pub use toolbar::*;
+pub use traffic_lights::*;
+pub use workspace::*;
