+++ crates/ui2/src/to_extract/assistant_panel.rs
