+++ crates/ui2/src/to_extract/chat_panel.rs
