+++ crates/ui2/src/to_extract/copilot.rs
