+++ crates/ui2/src/to_extract/notifications_panel.rs
