+++ crates/ui2/src/to_extract/project_panel.rs
