+++ crates/ui2/src/to_extract/status_bar.rs
