+++ crates/ui2/src/to_extract/terminal.rs
