+++ crates/ui2/src/to_extract/traffic_lights.rs
